feat(requester): override APIRequestFailed.__str__ to exclude request params
Header, which contains credentials such as bot token, is part of the
request params.
Previously, when HTTPRequester.APIRequestFailed is raised, the entire
request param is printed out, or included in logs, causing credentials
being leaked out.
This commit provides HTTPRequester.APIRequestFailed with a overriden
__str__ implementation in which params is not included by default.
In cases params is needed, repr() can be used.
This commit also provides a docstring with similar information.